Add files via upload
Notes: 
diseases+count contains the titles followed by the misdiagnoses for each title
at the end of the file, there are counts of how many times several different cases of patterns in the title occurred (ex. 1 disease misdiagnosed as 1 disease, 2 diseases misdiagnosed as 1 disease, 3 diseases misdiagnosed as 2 diseases, etc), and below that are the counts of each specific diseases pair.

there were many cases of 0 diseases misdiagnosed as _ diseases or _ diseases misdiagnosed as 0 diseases. the zero-zero file contains all the titles where this is the case and what the quickumls is recognizing in the title